+++ apps/admin_panel/assets/src/utils/url.js
